add casts to avoid warnings on Windows
git-svn-id: https://svn.r-project.org/R/trunk@65408 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/eval.c
@@ -223,7 +223,8 @@ static void doprof(int sig) /* sig is ignored in Windows */
     {
         get_current_mem(&smallv, &bigv, &nodes);
         if ((len = strlen(buf)) < PROFLINEMAX)
-            snprintf(buf + len, PROFBUFSIZ - len, ":%ld:%ld:%ld:%ld:", smallv, bigv, nodes, get_duplicate_counter());
+            snprintf(buf + len, PROFBUFSIZ - len, ":%lu:%lu:%lu:%lu:", (unsigned long)smallv, (unsigned long)bigv,
+                     (unsigned long)nodes, get_duplicate_counter());
         reset_duplicate_counter();
     }
